requirements.txt
hydra-core
omegaconf
litellm
python-dotenv
openai
anthropic
google-generativeai
cohere
tqdm